resovled some of the fee calculations in testing

+++ contracts/implementation/LeveragedPool.sol
@@ -68,7 +68,7 @@ contract LeveragedPool is ILeveragedPool, AccessControl, Initializable {
     int256 _firstPrice,
     uint32 _updateInterval,
     uint32 _frontRunningInterval,
-    uint16 _fee,
+    uint40 _fee,
     uint16 _leverageAmount,
     address _feeAddress,
     address _quoteToken
